Version 1.0.1209-b2 commit (Siraj) - New event logic with event names - Chat control interaction events with new event logic

+++ src/app/modules/t-widgets/tw-collections/tw-custom/tw-custom.component.ts
@@ -100,11 +100,13 @@ export class TwCustomComponent extends TWidgetWrapper implements OnInit, OnDestr
                 );
 
                 try {
-                    // listen to widget close event
-                    this.oinWidget.onunload = () => {
-                        // destroy the widget
-                        this._aotWidgetService.destroyWidget(this.data.ID);
-                    };
+                    if (this.oinWidget) {
+                        // listen to widget close event
+                        this.oinWidget.onunload = () => {
+                            // destroy the widget
+                            this._aotWidgetService.destroyWidget(this.data.ID);
+                        };
+                    }
                 } catch (error) {
                     console.error(error);
                 }
